fix msvc warnings-as-errors

+++ code/Common/FileSystemFilter.h
@@ -300,7 +300,7 @@ private:
 
         const char separator = getOsSeparator();
         for (it = in.begin(); it != in.end(); ++it) {
-            int remaining = std::distance(in.end(), it);
+            int remaining = (int)std::distance(in.end(), it);
             // Exclude :// and \\, which remain untouched.
             // https://sourceforge.net/tracker/?func=detail&aid=3031725&group_id=226462&atid=1067632
             if (remaining >= 3 && !strncmp(&*it, "://", 3 )) {
